Fix foreign key constraint error: Change Employee.CompanyId from Long to Integer to match Companies.companyId type

+++ src/main/java/com/application/employee/service/entities/Employee.java
@@ -57,7 +57,7 @@ public class Employee {
     private Companies company;
 
     @Column(name = "COMPANY_ID", insertable = false, updatable = false)
-    private Long CompanyId;
+    private Integer CompanyId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "ROLE")
